Trying optimization

+++ Assignment1/framework1/triangle_rasterization/trirast.h
@@ -3,10 +3,10 @@
 
 #include "types.h"
 
-void    draw_triangle(float x0, float y0, float x1, float y1,
-            float x2, float y2, byte r, byte g, byte b);
-void    draw_triangle_optimized(float x0, float y0, float x1, float y1,
-            float x2, float y2, byte r, byte g, byte b);
+void    draw_triangle(double x0, double y0, double x1, double y1,
+            double x2, double y2, byte r, byte g, byte b);
+void    draw_triangle_optimized(double x0, double y0, double x1, double y1,
+            double x2, double y2, byte r, byte g, byte b);
 
 int getMin(int a, int b, int c);
 int getMax(int a, int b, int c);
